scanner_tool/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scanner_tool"
version = "0.1.0"
description = "PC 端 3D 场景处理工具"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "numpy>=1.20",
    "scipy>=1.6.0",
    "pillow>=8.1.0",
    "opencv-python>=4.1",
    "open3d>=0.12.0",
    "scikit-video>=1.1.11",
    "torch>=2.0",
    "torchvision>=0.15",
    "tqdm",
    "rich",
    "matplotlib",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "hypothesis",
]

[project.scripts]
scanner-tool = "scanner_tool.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
